requirements.txt
# HTTP Requests and URL handling
requests==2.31.0
urllib3==2.0.7

# HTML/XML Parsing
beautifulsoup4==4.12.2
lxml==5.1.0

# Content Extraction
trafilatura==1.7.0
html2text==2020.1.16
readability-lxml==0.8.1

# Text Processing
langdetect==1.0.9
nltk==3.8.1

# Utilities
fake-useragent==1.4.0
python-dotenv==1.0.0
retry==0.9.2

# Caching (optional)
requests-cache==1.1.1

# Logging
coloredlogs==15.0.1

# Token counting for LLM chunking
tiktoken==0.5.2
